fix: hotfix timebar not showing on all channels & slices
Before this commit, the timebar was not showing correctly on
all channels & slices. The timebar for t = 0 would show at z = 0
only, the timebar for t = 1 would show at z = 1 only, and so on...

This was due to a difference in stacks vs hyperstacks in the way
ImageJ displays overlays.